Added functionality to allow the user to create furniture and add it to the room they just created, also added a new utilities class.
The method to create furniture was added to the room class, this is called from the house after the room is created. This allows the user to add furniture to the room they just created.

As the loop within the house class would also use the method to see if the user wishes to end the loop, this method was moved to a new utlities class that can be accessed anywhere within the program.

Changed two methods that were in camelcase to pascal case.