Implemented a function to remove houses from the list.
Added in method to remove a house, the user is shown the list and selects a house from its number, this house gets located and then removed. Added in the option to the switch case in the main function.

Moved the code that calls the create room function to the function that creates houses. This was done to clean up the Main function.