Started the function to edit houses, currently you can change the name and number of a house. Moved some code into its own function for reuse, and gave the user the option to display the list of houses.
Added in a function to allow the user to edit houses, currently they have the option to change the name and number of a house.

Moved the loops that get a name and number off the user into there own functions. This was done as the edit house function will use the same loops to get a new name or number off the user.

Changed the code to find a house, added the code from the remove function that gives the user the option to select a house as the edit function would also use this code.

Added in the option to display the list of houses, so the user can view them at any point from the main menu.